fixed save_message
Version 2.0.62
General:
- updated addons copyright and media link

module updates:
- save_message | fixed

+++ save_message.py
@@ -1,10 +1,10 @@
-__version__ = (0, 0, 29)
+__version__ = (0, 0, 32)
 
 
 # ▄▀█ █▄ █ █▀█ █▄ █ █▀█ ▀▀█ █▀█ █ █ █▀
 # █▀█ █ ▀█ █▄█ █ ▀█ ▀▀█   █ ▀▀█ ▀▀█ ▄█
 #
-#           © Copyright 2023
+#           © Copyright 2024
 #
 #        developed by @anon97945
 #
@@ -15,8 +15,8 @@ __version__ = (0, 0, 29)
 # 🌐 https://www.gnu.org/licenses/gpl-3.0.html
 
 # meta developer: @apodiktum_modules
-# meta banner: https://t.me/file_dumbster/11
-# meta pic: https://t.me/file_dumbster/13
+# meta banner: https://t.me/apodiktum_dumpster/11
+# meta pic: https://t.me/apodiktum_dumpster/13
 
 # scope: hikka_only
 # scope: hikka_min 1.3.3
@@ -24,6 +24,7 @@ __version__ = (0, 0, 29)
 import logging
 
 from telethon.tl.types import Message
+from io import BytesIO
 
 from .. import loader, utils
 
@@ -107,11 +108,35 @@ class ApodiktumSaveMessageMod(loader.Module):
             )
         channel_id, msg_id = self.apo_lib.utils.get_ids_from_tglink(args)
         msgs = await self._client.get_messages(channel_id, ids=msg_id)
-        msgs = await self._client.send_message(self.tg_id, message=msgs)
-        await utils.answer(
-            message,
-            self.apo_lib.utils.get_str("done", self.all_strings, message),
-        )
+        try:
+            msgs = await msgs.forward_to(self.tg_id)
+        except Exception as exc:  # skipcq: PYL-W0703
+            if "You can't forward messages from a protected chat (caused by ForwardMessagesRequest)" in str(exc):
+                if not msgs.file:
+                    msgs = await self._client.send_message(self.tg_id, message=msgs)
+                    await utils.answer(
+                        message,
+                        self.apo_lib.utils.get_str("done", self.all_strings, message),
+                    )
+                    return
+                file = BytesIO()
+                caption = f"{utils.escape_html(msgs.text)}"
+                await self._client.download_file(msgs, file)
+                file.name = (
+                    msgs.file.name or f"{msgs.file.media.id}{msgs.file.ext}"
+                )
+                file.seek(0)
+                msgs = await self._client.send_file(
+                    self.tg_id,
+                    file,
+                    force_document=True,
+                    caption=caption,
+                )
+                await utils.answer(
+                    message,
+                    self.apo_lib.utils.get_str("done", self.all_strings, message),
+                )
+                return
 
     async def smhcmd(self, message: Message):
         """<messagelink> to forward message/media to current chat."""
@@ -127,9 +152,29 @@ class ApodiktumSaveMessageMod(loader.Module):
 
         channel_id, msg_id = self.apo_lib.utils.get_ids_from_tglink(args)
         msgs = await self._client.get_messages(channel_id, ids=msg_id)
-        msgs = await self._client.send_message(
-            utils.get_chat_id(message),
-            message=msgs,
-        )
-
-        await message.delete()
+        try:
+            msgs = await self._client.send_message(
+                utils.get_chat_id(message),
+                message=msgs,
+            )
+            await message.delete()
+        except Exception as exc:  # skipcq: PYL-W0703
+            if "You can't forward messages from a protected chat (caused by ForwardMessagesRequest)" in str(exc) or "You can't forward messages from a protected chat (caused by SendMediaRequest)" in str(exc):
+                if not msgs.file:
+                    await message.delete()
+                    return
+                file = BytesIO()
+                caption = f"{utils.escape_html(msgs.text)}"
+                await self._client.download_file(msgs, file)
+                file.name = (
+                    msgs.file.name or f"{msgs.file.media.id}{msgs.file.ext}"
+                )
+                file.seek(0)
+                msgs = await self._client.send_file(
+                    utils.get_chat_id(message),
+                    file,
+                    force_document=True,
+                    caption=caption,
+                )
+                await message.delete()
+                return
